add users

+++ MainApp/views.py
@@ -1,8 +1,9 @@
 from django.http import Http404
-from django.shortcuts import render, redirect
+from django.shortcuts import get_object_or_404, render, redirect
 from MainApp.forms import SnippetForm
 from MainApp.models import Snippet
-
+from django.core.exceptions import ObjectDoesNotExist
+from django.contrib import auth
 
 def index_page(request):
     context = {'pagename': 'PythonBin'}
@@ -23,7 +24,10 @@ def add_snippet_page(request):
     if request.method == "POST":
         form = SnippetForm(request.POST)
         if form.is_valid():
-            form.save()
+            snippet = form.save(commit=False)
+            if request.user.is_authenticated:
+                snippet.user = request.user
+                snippet.save()
             return redirect("snippets-list")  # GET /snippets/list
         return render(request,'pages/add_snippet.html', {'form': form})
 
@@ -47,17 +51,75 @@ def snippet_detail(request, snippet_id):
         context["snippet"] = snippet
         return render(request, "pages/snippet_detail.html", context)
     
-def delete_snipet(request, snippet_id):
-    snippet = Snippet.objects.get(pk=snippet_id)
-    snippet.delete()
+
+
+ 
+def snippet_edit(request, snippet_id):
+    context = {'pagename': 'Редактирование сниппета'}
+    try:
+        snippet = Snippet.objects.get(pk=snippet_id)
+    except ObjectDoesNotExist:
+        raise Http404
+    
+    # Variant 1 с помощью SnippetForm
+    # if request.method == "GET":
+    #     form = SnippetForm(instance=snippet)
+    #     return render(request,"pages/add_snippet.html", {"form": form})
+    
+    # Variant 2
+    # Хотим получить страницу с данными сниппета
+    if request.method == "GET":
+        context = {
+            "snippet": snippet,
+            "type": "edit"
+        }
+        return render(request, "pages/snippet_detail.html", context)
+    
+    # Получаем данные из формы и на их основе меняем значения Сниппета
+    if request.method == "POST":
+        data_form = request.POST
+        snippet.name = data_form["name"]
+        snippet.code = data_form["code"]
+        snippet.save()
+        return redirect("snippets-list")
+
+
+
+
+
+
+# def snippet_delete(request, snippet_id):
+#     snippet = Snippet.objects.get(id=snippet_id)
+#     snippet.delete()
+#     return redirect('snippets-list')
+
+# def snippet_delete(request, snippet_id):
+#     snippet = get_object_or_404(Snippet, id = snippet_id)
+#     snippet.delete()
+#     return redirect('snippets-list')
+
+def snippet_delete(request, snippet_id):
+    if request.method == "GET" or request.method == "POST":
+        snippet = get_object_or_404(Snippet, id=snippet_id)
+        snippet.delete()
     return redirect('snippets-list')
 
-def edit_snipet(request, snippet_id, name, lang, code, date):
-    snippet = Snippet.objects.get(id=snippet_id)
-    snippet.name = name
-    snippet.lang = lang
-    snippet.code = code
-    snippet.creation_date = date
-    snippet.save(update_fields=["name", "lang", "code", "creation_date"])
-    return redirect('snippets-list')
- 
+
+def login(request):
+    if request.method == 'POST':
+        username = request.POST.get("username")
+        password = request.POST.get("password")
+        # print("username =", username)
+        # print("password =", password)
+        user = auth.authenticate(request, username=username, password=password)
+        if user is not None:
+            auth.login(request, user)
+        else:
+            # Return error message
+            pass
+    return redirect('home')
+
+
+def logout(request):
+    auth.logout(request)
+    return redirect("home")
